Move pure test code into test hierarchy
